Disabling the default recursor service
It stops/disables the service on Debian platforms

It disables the service on RHEL, no need to stop
the recursor in this case because its not launched 
in the first place. 

This way we remove only_if ward on the pidfile
which was preventing chef to stop/disable the 
recursor on Ubuntu. 

It’s also a cleaner solution that leaves to the
underlying os to do what’s required