Refactor resources to match packaging configurations (#80)
* TDD sysvinit_name

Add a test for sysvinit_name method with correct 
name.

Correct naming scheme for that method for good,
reflecting when parameter is nil (basic case) and
when its named as well.

* TDD sysvinit_name for recursor

Doing the same changes for recursor instances on
sysvinit

* Adding a note on future removal

* Fixing spec by using regular inspec DSL

* Check on empty? method for strings

* Rename to better wrap around pdns defaults

* Improving test cases by adding a ‘’ test

* Adding tests for instance-config methods

* Virtual hosting on recursor

Adapting debian init script from authoritative to
support virtual hosting on recursor

* Updating spec to match the new changes

* Recursor helpers update

- Checking for empty?
- powerdns_recursor assumes a ‘recursor’ prefix
for config files other than the default which
should be ‘pdns-recursor.conf’

* Adapting resource to use config helper

* Updating init resource

- Stop pdns-recursor only when we have a 
default, base-case install (when calling 
pdns_recursor_serv ice with ‘’). To allow updating 
a configuration change of this instance.

- Updating pdns-recursor init script to allow
multiple recursor instances by using a symlink

* Updating inspec test and test-recipe

To match the edge case of having a default (“”)
resource.

* Deprecated code

- It was used for tests/debug

* Improved regexp to catch everything

* WIP

* Basic case working

* Adding pidfile option to the daemon lsb script

In order to support multiple instances. The lsb 
daemon script requires —-pidfile option

* Improve matching on regular expressions

To better match both pdns/pdns-recursor on 
debian/rhel based distros
Also converting everything to regexp classes

* Removing Centos 6.7 platform

* Fixes for systemd helpers for names plus tests

* Spacing 🚨

* This is a particular case

When we use the default recursor, we want to
stop it before starting it again for the first
time since it grab config changes.

* Remove recipe tests

These are covered via integration tests. Additional coverage is still
needed in the library helpers.

* Cleanup spec helper

* Move this folder back one tier

It's assumed we're doing unit tests here.

* Properly name the library helpers spec

* Cleanup naming and sourcing

* Write the init out to the correct location

* I never meta test I couldn't meta test :D

No seriously, this helps deal with the nuances between sysvinit and
systemd process running styles.

* Make sure we use the correct config directory

This caused breakage in the RHEL family tests which use entirely
different directories.

* Don't go metaling with service state

* Break up and re-order the authoritative install

* Use the same matching trick here too

* This is a bit overkill

It also buries the test suite with indirection :(

* Sorry :cop: :(

* This is implicit

* Shuffle the order a little bit

Also don't be as aggressive about service notification

* Fix a missing property and document it

* Using a data structure to contain the information

* Reorder

* Fix, replace symbols per strings on hash

* Access the hash properly

* Removing comments

* Sorting out backends

* Covering every helper with tests

* Updating default_authoritative_config_directory

It was called in an inconsistent way between
resources.

I avoid calling a cookbook library in a 
test cookbook

* Remove backend resource

This does nothing more than wrap the package resource. It is not
maintainable or sustainable for us to keep a massive hash of possible
backends especially if they change from underneath us or get
inconsistent between platforms and versions.

* Cleanup the documentation a little more

* Sup :cop:

* Re-do the breaking change notes for upcoming 4.x

* Rename helpers (we don't need pdnspdnspdnspdns)

* Upgrade to Chef 12.14+

* Move and rename resources

The names are implicit via provides

* I have a feeling this isn't an issue anymore

* Adding a function to check if there’s hyphens

* Adding a function to check if there’s hyphens

* raise_an exception on names with hyphens

* More tests and corrections

- Adding tests for names with hyphens
- Correcting some calls in expectation that were
copied from other tests and were calling to the
wrong function

* Looking for hyphens in config names

* Adding a second test for hyphens

* Use property callbacks to prevent hyphens

* Remove hyphen helper

This is replaced by validation callbacks in the resources

* This is no longer required

ChefSpec as of the ChefDK 2.0 release now automatically generates them.

* Test for a raised validation exception

* Add additional test coverage

* Switch to do/end blocks

* This is no lonegr a requirement with Chef 12.14

* Use an absolute path for safety

* Remove nilclass option and make source optional

This re-works and simplifies the logic to work only if a custom source
has been set.

* Remove nil as an option here

It must be defined for Chef to be able to find the template source file
and which cookbook it should look in for it.

* Remove nilclass as an option here

Chef won't know where to look and what to look for if it is not defined.

* Set to an absolute path for safety

* Pass in the entire set of variables

* Reformat documentation and note missing variables

* Pass variables hash directly instead of burying it