Fix config permissions
The default permissions were too strict. We didn't want the config files
writeable by PowerDNS becuase it's not necessary for operation but they
need to be readable x.x

This splits the difference by not making them world readable but making
the pdns user the group.

This might be a touch too far and we may want to 644 and make the user
and group that set by the resource. Feedback welcome.

Signed-off-by: David Aronsohn <WagThatTail@Me.com>