Fix config permissions (#90)
The default permissions were too strict. We didn't want the config files
writeable by PowerDNS because it's not necessary for operation but they
need to be readable x.x

This splits the difference by not making them world readable but making
the pdns user's primary group have access to the file.